+++ modules/ensemble/lib/framework/apiproviders/firestore/firestore_api_provider.dart
@@ -274,7 +274,6 @@ class FirestoreAPIProvider extends APIProvider with LiveAPIProvider {
       'message': 'Subscribed to API',
       'documents': []
     };
-    ;
     return FirestoreResponse(
       apiState: APIState.success,
       body: body,