+++ modules/ensemble/lib/screen_controller.dart
@@ -2,6 +2,7 @@
 
 import 'dart:async';
 
+import 'package:ensemble/action/face_camera_actions.dart';
 import 'package:ensemble/action/navigation_action.dart';
 import 'package:ensemble/action/upload_files_action.dart';
 import 'package:ensemble/ensemble.dart';
@@ -15,6 +16,7 @@ import 'package:ensemble/framework/error_handling.dart';
 import 'package:ensemble/framework/event.dart';
 import 'package:ensemble/framework/scope.dart';
 import 'package:ensemble/framework/stub/camera_manager.dart';
+import 'package:ensemble/framework/stub/face_camera_manager.dart';
 import 'package:ensemble/framework/theme/theme_loader.dart';
 import 'package:ensemble/framework/theme_manager.dart';
 import 'package:ensemble/framework/view/data_scope_widget.dart';
@@ -262,6 +264,8 @@ class ScreenController {
       }
     } else if (action is ShowCameraAction) {
       GetIt.I<CameraManager>().openCamera(context, action, scopeManager);
+    } else if (action is ShowFaceCameraAction) {
+      GetIt.I<FaceCameraManager>().openFaceCamera(context, action, scopeManager);
     } else if (action is StartTimerAction) {
       // validate
       bool isRepeat = action.isRepeat(scopeManager.dataContext);