+++ modules/ensemble/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.2.34
+
+ - **FEAT**(remote_config): add Firebase Remote Config integration. ([906f0133](https://github.com/ensembleUI/ensemble/commit/906f013322dcda45a1740db24b5e21f63ea372e5))
+
+## 1.2.33
+
+ - **FEAT**(face_camera): separate face detection from ensemble_camera into separate module. ([9857c3cc](https://github.com/ensembleUI/ensemble/commit/9857c3cc3048f2c8879feeb07d77b084181010cd))
+
+## 1.2.32
+
+ - Upgrading ensemble version to 1.2.32
+
 ## 1.2.31
 
  - **FEAT**: add spans support to Text widget for inline rich text and widget embedding. ([69f92b19](https://github.com/ensembleUI/ensemble/commit/69f92b19502c9b7db8248123e192f808b64d1869))